more debug nan losses

+++ predictor.py
@@ -311,9 +311,10 @@ class Predictor(OldPredictor):
         print(subtitle_mask)
         print("SUBTITLE_MASK_SHAOPE:", subtitle_mask.shape)
         print("LOCAL_VP_SHAPE:", local_vp.shape)
-        local_vp = local_vp.reshape(-1,)
+        local_vp = local_vp.reshape(-1,).astype(np.float32)
+        subtitle_mask = subtitle_mask.astype(np.float32) 
         print("LOCAL_VP_SHAPE:", local_vp.shape)
-        print(np.unique(subtitle_mask)) 
+        print(np.unique(subtitle_mask))
         for i in np.arange(0, head_room):
             temp = log_loss(
                     subtitle_mask,
